+++ CHANGES.rst
@@ -4,6 +4,7 @@ Changelog
 2.7.0 (unreleased)
 ------------------
 
+- #67 Rename "Recover" transtion to "Retrieve" to align with ISO 17025/15189
 - #66 Add visual warning for samples approaching retention expiration
 - #65 Add a "Past Retention" filter in samples listing
 - #63 Result type-specific controls in retention rules settings